macOS compatibility fixes
