+++ osquery/tables/system/darwin/tests/xattr_utils_tests.cpp
@@ -18,16 +18,17 @@
 #include <osquery/filesystem.h>
 #include <osquery/logger.h>
 
+namespace fs = boost::filesystem;
+
 namespace osquery {
+Status readSpecialExtendedAttribute(
+    std::vector<std::pair<std::string, std::string>>& output,
+    const std::string& path,
+    const std::string& name);
+
 namespace tables {
-
-std::vector<std::string> parseExtendedAttributeList(const std::string& path);
-void parseQuarantineFile(QueryData& results, const std::string& path);
-void parseWhereFrom(QueryData& results, const std::string& path);
-void getFileData(QueryData& results, const std::string& path);
-
-const std::string kQuarantineKey = "com.apple.quarantine";
-const unsigned char kQuarantineValue[] = {
+const std::string kQuarantineXattr = "com.apple.quarantine";
+const std::uint8_t kQuarantineXattrValue[] = {
     0x30, 0x30, 0x30, 0x31, 0x3B, 0x35, 0x35, 0x34, 0x39, 0x34, 0x64, 0x65,
     0x33, 0x3B, 0x47, 0x6F, 0x6F, 0x67, 0x6C, 0x65, 0x5C, 0x78, 0x32, 0x30,
     0x43, 0x68, 0x72, 0x6F, 0x6D, 0x65, 0x3B, 0x36, 0x41, 0x30, 0x45, 0x43,
@@ -35,111 +36,85 @@ const unsigned char kQuarantineValue[] = {
     0x38, 0x2D, 0x41, 0x33, 0x31, 0x35, 0x2D, 0x35, 0x38, 0x36, 0x42, 0x41,
     0x36, 0x35, 0x46, 0x38, 0x46, 0x37, 0x37};
 
-const std::string kMetedataKey = "com.apple.metadata:kMDItemWhereFroms";
-const unsigned char kMetadataWhereFromValue[] = {
-    0x62, 0x70, 0x6C, 0x69, 0x73, 0x74, 0x30, 0x30, 0xA2, 0x01, 0x02, 0x5F,
-    0x10, 0x3D, 0x68, 0x74, 0x74, 0x70, 0x73, 0x3A, 0x2F, 0x2F, 0x64, 0x6C,
-    0x2E, 0x67, 0x6F, 0x6F, 0x67, 0x6C, 0x65, 0x2E, 0x63, 0x6F, 0x6D, 0x2F,
-    0x63, 0x68, 0x72, 0x6F, 0x6D, 0x65, 0x2F, 0x6D, 0x61, 0x63, 0x2F, 0x73,
-    0x74, 0x61, 0x62, 0x6C, 0x65, 0x2F, 0x47, 0x47, 0x52, 0x4F, 0x2F, 0x67,
-    0x6F, 0x6F, 0x67, 0x6C, 0x65, 0x63, 0x68, 0x72, 0x6F, 0x6D, 0x65, 0x2E,
-    0x64, 0x6D, 0x67, 0x5F, 0x10, 0x40, 0x68, 0x74, 0x74, 0x70, 0x73, 0x3A,
-    0x2F, 0x2F, 0x77, 0x77, 0x77, 0x2E, 0x67, 0x6F, 0x6F, 0x67, 0x6C, 0x65,
-    0x2E, 0x63, 0x6F, 0x6D, 0x2F, 0x63, 0x68, 0x72, 0x6F, 0x6D, 0x65, 0x2F,
-    0x62, 0x72, 0x6F, 0x77, 0x73, 0x65, 0x72, 0x2F, 0x74, 0x68, 0x61, 0x6E,
-    0x6B, 0x79, 0x6F, 0x75, 0x2E, 0x68, 0x74, 0x6D, 0x6C, 0x3F, 0x70, 0x6C,
-    0x61, 0x74, 0x66, 0x6F, 0x72, 0x6D, 0x3D, 0x6D, 0x61, 0x63, 0x08, 0x0B,
-    0x4B, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x01, 0x00, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x8E};
+class MacOSExtendedAttrTests : public testing::Test {};
 
-const std::string kFsckKey = "com.apple.diskimages.fsck";
-const unsigned char kFsckValue[] = {0x38, 0xFF, 0xBE, 0x7E, 0xEF, 0xBE, 0x27,
-                                    0xC0, 0x7B, 0x00, 0x6C, 0x88, 0xB2, 0xA1,
-                                    0x50, 0x4D, 0x16, 0x70, 0x6C, 0x7A};
+TEST_F(MacOSExtendedAttrTests, test_invalid_special_xattr) {
+  auto test_file_path =
+      (fs::temp_directory_path() / fs::unique_path()).string();
 
-const std::string kDiskImagefsckBase64 = "OP++fu++J8B7AGyIsqFQTRZwbHo=";
-
-void removexattrs(const std::string& path) {
-  auto xattrs = parseExtendedAttributeList(path);
-  for (const auto& xattr : xattrs) {
-    removexattr(path.c_str(), xattr.c_str(), XATTR_NOFOLLOW);
+  {
+    std::fstream test_file(test_file_path, std::ios::out);
+    EXPECT_EQ(!test_file, false);
   }
+
+  const char dummy_string[] = "test";
+  EXPECT_EQ(setxattr(test_file_path.c_str(),
+                     dummy_string,
+                     dummy_string,
+                     sizeof(dummy_string),
+                     0,
+                     XATTR_NOFOLLOW),
+            0);
+
+  // The function should not attempt to expand this attribute
+  std::vector<std::pair<std::string, std::string>> expanded_attributes;
+  auto status = readSpecialExtendedAttribute(
+      expanded_attributes, test_file_path, dummy_string);
+  EXPECT_TRUE(status.ok());
+
+  EXPECT_EQ(expanded_attributes.size(), 0U);
+
+  fs::remove(test_file_path);
 }
 
-void setxattrs(const std::string& path) {
-  setxattr(path.c_str(),
-           kQuarantineKey.c_str(),
-           (void*)kQuarantineValue,
-           sizeof(kQuarantineValue),
-           0,
-           XATTR_NOFOLLOW);
-  setxattr(path.c_str(),
-           kMetedataKey.c_str(),
-           (void*)kMetadataWhereFromValue,
-           sizeof(kMetadataWhereFromValue),
-           0,
-           XATTR_NOFOLLOW);
-  setxattr(path.c_str(),
-           kFsckKey.c_str(),
-           (void*)kFsckValue,
-           sizeof(kFsckValue),
-           0,
-           XATTR_NOFOLLOW);
-  // insert arbitrary xattr
-  const std::string key = "foobar";
-  const unsigned char val[] = {0x62, 0x61, 0x7A}; // baz
-  setxattr(
-      path.c_str(), key.c_str(), (void*)val, sizeof(val), 0, XATTR_NOFOLLOW);
-}
+TEST_F(MacOSExtendedAttrTests, test_quarantine_xattrs) {
+  auto test_file_path =
+      (fs::temp_directory_path() / fs::unique_path()).string();
 
-class ExtendedAttributesTests : public testing::Test {
- protected:
-  virtual void SetUp() {
-    removexattrs(kTestFilePath);
-    setxattrs(kTestFilePath);
+  {
+    std::fstream test_file(test_file_path, std::ios::out);
+    EXPECT_EQ(!test_file, false);
   }
 
-  virtual void TearDown() {
-    removexattrs(kTestFilePath);
-  }
+  EXPECT_EQ(setxattr(test_file_path.c_str(),
+                     kQuarantineXattr.c_str(),
+                     kQuarantineXattrValue,
+                     sizeof(kQuarantineXattrValue),
+                     0,
+                     XATTR_NOFOLLOW),
+            0);
 
-  const std::string kTestFilePath = kTestDataPath + "test_xattrs.txt";
-  const std::string kTestFileDir =
-      boost::filesystem::path(kTestFilePath).parent_path().string();
-};
-
-TEST_F(ExtendedAttributesTests, test_extended_attributes) {
-  QueryData results;
-  getFileData(results, kTestFilePath);
+  std::vector<std::pair<std::string, std::string>> expanded_attributes;
+  auto status = readSpecialExtendedAttribute(
+      expanded_attributes, test_file_path, kQuarantineXattr);
+  EXPECT_TRUE(status.ok());
 
   const std::map<std::string, std::string> expected = {
       {"quarantine_agent", "Google Chrome"},
-      {"quarantine_type", "LSQuarantineTypeWebDownload"},
-      {"quarantine_timestamp", "1430867421"},
       {"quarantine_event_id", "6A0EC92D-1882-4358-A315-586BA65F8F77"},
-      {"quarantine_data_url",
-       "https://dl.google.com/chrome/mac/stable/GGRO/googlechrome.dmg"},
-      {"quarantine_origin_url",
-       "https://www.google.com/chrome/browser/thankyou.html?platform=mac"},
-      {"foobar", "baz"},
-      {kFsckKey, kDiskImagefsckBase64}};
+      {"quarantine_timestamp", "1430867427"}};
 
-  for (auto row : results) {
-    for (auto elems : row) {
-      auto key = elems.first;
-      auto value = elems.second;
-      if (expected.count(key) > 0) {
-        if (key == "quarantime_timestamp") {
-          long timeExpected = std::stol(expected.at("quarantine_timestamp"));
-          long actual = std::stol(value);
-          EXPECT_GE(timeExpected, actual);
-        } else {
-          EXPECT_EQ(expected.at(key), value);
-        }
-      }
+  size_t matching_entries = 0U;
+
+  for (const auto& p : expanded_attributes) {
+    const auto& name = p.first;
+    const auto& value = p.second;
+
+    auto it = expected.find(name);
+    if (it == expected.end()) {
+      continue;
     }
+
+    if (it->second != value) {
+      continue;
+    }
+
+    matching_entries++;
   }
+
+  EXPECT_EQ(matching_entries, expected.size());
+
+  fs::remove(test_file_path);
 }
 } // namespace tables
 } // namespace osquery
