--- tools/tests/test_xattrs.txt
@@ -1,1 +1,0 @@
-This file is used to test extended attributes.
